Bring Public fork up-to-date
This commit is 219mb in size from the overwrite of changes.
All this involves a culmination of many months of hardwork and reverse engineering.
With this comment introduces many new things.
FIrstly, lays foundations for all other DME Game Server packets types.
Implements Medius File Services properly.
Implements Buddy Invitations properly.
Implements Peer to Peer, needs a bit more work but functional.
Beginnings of Universes over DB
Beginnings of BWPS, GHS, MUM, MMS sub servers.
Beginnings of Anticheat and HTTPService for the servers.
Beginnings of SVO.
MUIS now supports older PS2 titles prior to v2.10

+++ RT.Models/Lobby/MediusGameListResponse.cs
@@ -18,7 +18,7 @@ namespace RT.Models
         public string GameName;
         public MediusWorldStatus WorldStatus;
         public MediusGameHostType GameHostType;
-        public ushort PlayerCount;
+        public int PlayerCount;
         public bool EndOfList;
 
         public override void Deserialize(Server.Common.Stream.MessageReader reader)
@@ -36,7 +36,7 @@ namespace RT.Models
             GameName = reader.ReadString(Constants.GAMENAME_MAXLEN);
             WorldStatus = reader.Read<MediusWorldStatus>();
             GameHostType = reader.Read<MediusGameHostType>();
-            PlayerCount = reader.ReadUInt16();
+            PlayerCount = reader.ReadInt32();
             EndOfList = reader.ReadBoolean();
             reader.ReadBytes(3);
         }
